refactor(metrics): drop _context shim, import ScoringContext from utils
